Cherry-pick from production based-branch
