+++ src/main/java/beam/analysis/plots/GraphSurgePricing.java
@@ -1,9 +1,7 @@
 package beam.analysis.plots;
 
+import beam.agentsim.agents.ridehail.AdaptiveRideHailSurgePricingManager;
 import beam.agentsim.agents.ridehail.RideHailSurgePricingManager;
-import beam.agentsim.agents.ridehail.SurgePriceBin;
-import beam.sim.OutputDataDescription;
-import beam.utils.OutputDataDescriptor;
 import com.google.inject.Inject;
 import org.jfree.chart.ChartFactory;
 import org.jfree.chart.JFreeChart;
@@ -21,6 +19,7 @@ import org.matsim.core.controler.listener.IterationEndsListener;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import scala.collection.Iterator;
+import beam.agentsim.agents.ridehail.RideHailSurgePricingManager.SurgePriceBin;
 import scala.collection.mutable.ArrayBuffer;
 
 import java.io.BufferedWriter;
@@ -60,11 +59,11 @@ public class GraphSurgePricing implements ControlerListener, IterationEndsListen
     private String surgePricingAndRevenueWithTaz = "";
     private String revenueGraphImageFile = "";
     private String revenueCsvFileName = "";
-    private final RideHailSurgePricingManager surgePricingManager;
+    private final AdaptiveRideHailSurgePricingManager surgePricingManager;
     private final boolean writeGraph;
 
     @Inject
-    public GraphSurgePricing(RideHailSurgePricingManager surgePricingManager) {
+    public GraphSurgePricing(AdaptiveRideHailSurgePricingManager surgePricingManager) {
         this.surgePricingManager = surgePricingManager;
         noOfCategories = this.surgePricingManager.numberOfCategories();
 
@@ -164,7 +163,7 @@ public class GraphSurgePricing implements ControlerListener, IterationEndsListen
         }
     }
 
-    private void processSurgePriceBinsMap(RideHailSurgePricingManager surgePricingManager) {
+    private void processSurgePriceBinsMap(AdaptiveRideHailSurgePricingManager surgePricingManager) {
 
         scala.collection.immutable.Map<String, scala.collection.mutable.ArrayBuffer<SurgePriceBin>> surgePriceBinsMap = surgePricingManager.surgePriceBins();
         Iterator mapIter = surgePriceBinsMap.keysIterator();
@@ -179,7 +178,7 @@ public class GraphSurgePricing implements ControlerListener, IterationEndsListen
             double[][] _tazDataset = new double[2][numberOfTimeBins];
 
             for (int i = 0; iter.hasNext(); i++) {
-                SurgePriceBin bin = (SurgePriceBin) iter.next();
+                SurgePriceBin bin = (RideHailSurgePricingManager.SurgePriceBin) iter.next();
 
                 double price = bin.currentIterationSurgePriceLevel();
                 double revenue = bin.currentIterationRevenue();
@@ -194,7 +193,7 @@ public class GraphSurgePricing implements ControlerListener, IterationEndsListen
         }
     }
 
-    private void processBin(int binNumber, SurgePriceBin surgePriceBin) {
+    private void processBin(int binNumber, RideHailSurgePricingManager.SurgePriceBin surgePriceBin) {
 
         double revenue = surgePriceBin.currentIterationRevenue();
         revenueDataSet[binNumber] += revenue;
@@ -420,8 +419,8 @@ public class GraphSurgePricing implements ControlerListener, IterationEndsListen
 
                     double[] priceLevels = dataset[j];
 
-                    for(int i = 0; i < priceLevels.length; i++) {
-                        out.write(strFormat + "," + getRoundedNumber(priceLevels[i]) + "," + (i+1));
+                    for (int i = 0; i < priceLevels.length; i++) {
+                        out.write(strFormat + "," + getRoundedNumber(priceLevels[i]) + "," + (i + 1));
                         out.newLine();
                     }
                 }
@@ -431,8 +430,8 @@ public class GraphSurgePricing implements ControlerListener, IterationEndsListen
 
                     double[] priceLevels = dataset[j];
 
-                    for(int i = 0; i < priceLevels.length; i++) {
-                        out.write(categoriesList.get(j) + "," + getRoundedNumber(priceLevels[i]) + "," + (i+1));
+                    for (int i = 0; i < priceLevels.length; i++) {
+                        out.write(categoriesList.get(j) + "," + getRoundedNumber(priceLevels[i]) + "," + (i + 1));
                         out.newLine();
                     }
                 }
@@ -456,7 +455,7 @@ public class GraphSurgePricing implements ControlerListener, IterationEndsListen
                 double[] priceLevels = data[0];
                 double[] revenues = data[1];
 
-                for(int i = 0; i < priceLevels.length; i++) {
+                for (int i = 0; i < priceLevels.length; i++) {
                     out.write(tazId + ",pricelevel," + getRoundedNumber(priceLevels[i]) + "," + (i + 1));
                     out.newLine();
 
@@ -478,7 +477,7 @@ public class GraphSurgePricing implements ControlerListener, IterationEndsListen
             out.write("Revenue,Hour");
             out.newLine();
 
-            for(int i=0; i<revenueDataSet.length; i++){
+            for (int i = 0; i < revenueDataSet.length; i++) {
 
                 out.write(getRoundedNumber(revenueDataSet[i]) + "," + (i));
                 out.newLine();
