+++ src/main/java/beam/analysis/plots/RideHailRevenueAnalysis.java
@@ -1,6 +1,6 @@
 package beam.analysis.plots;
 
-import beam.agentsim.agents.ridehail.RideHailSurgePricingManager;
+import beam.agentsim.agents.ridehail.AdaptiveRideHailSurgePricingManager;
 import beam.analysis.plots.modality.RideHailDistanceRowModel;
 import com.google.inject.Inject;
 import org.jfree.chart.ChartFactory;
@@ -23,13 +23,13 @@ import static beam.analysis.AnalysisCollector.rideHailRevenueAnalytics;
 
 public class RideHailRevenueAnalysis implements ControlerListener, IterationEndsListener {
 
-    private RideHailSurgePricingManager surgePricingManager;
+    private AdaptiveRideHailSurgePricingManager surgePricingManager;
     static String fileBaseName = "rideHailRevenue";
 
     private OutputDirectoryHierarchy outputDirectoryHiearchy;
 
     @Inject
-    public RideHailRevenueAnalysis(RideHailSurgePricingManager surgePricingManager) {
+    public RideHailRevenueAnalysis(AdaptiveRideHailSurgePricingManager surgePricingManager) {
         this.surgePricingManager = surgePricingManager;
     }
