+++ src/main/scala/beam/sim/BeamMobsim.scala
@@ -8,7 +8,7 @@ import akka.pattern.ask
 import akka.util.Timeout
 import beam.agentsim.agents.BeamAgent.Finish
 import beam.agentsim.agents.ridehail.RideHailManager.{BufferedRideHailRequestsTrigger, RideHailRepositioningTrigger}
-import beam.agentsim.agents.ridehail.{RideHailIterationHistory, RideHailManager, RideHailSurgePricingManager}
+import beam.agentsim.agents.ridehail.{RideHailIterationHistory, RideHailManager, AdaptiveRideHailSurgePricingManager}
 import beam.agentsim.agents.vehicles.BeamVehicleType
 import beam.agentsim.agents.{BeamAgent, InitializeTrigger, Population, TransitSystem}
 import beam.agentsim.infrastructure.ZonalParkingManager
@@ -37,20 +37,20 @@ import scala.concurrent.Await
 import scala.concurrent.duration._
 
 class BeamMobsim @Inject()(
-  val beamServices: BeamServices,
-  val beamScenario: BeamScenario,
-  val transportNetwork: TransportNetwork,
-  val tollCalculator: TollCalculator,
-  val scenario: Scenario,
-  val eventsManager: EventsManager,
-  val actorSystem: ActorSystem,
-  val rideHailSurgePricingManager: RideHailSurgePricingManager,
-  val rideHailIterationHistory: RideHailIterationHistory,
-  val routeHistory: RouteHistory,
-  val beamSkimmer: BeamSkimmer,
-  val travelTimeObserved: TravelTimeObserved,
-  val geo: GeoUtils,
-  val networkHelper: NetworkHelper
+                            val beamServices: BeamServices,
+                            val beamScenario: BeamScenario,
+                            val transportNetwork: TransportNetwork,
+                            val tollCalculator: TollCalculator,
+                            val scenario: Scenario,
+                            val eventsManager: EventsManager,
+                            val actorSystem: ActorSystem,
+                            val rideHailSurgePricingManager: AdaptiveRideHailSurgePricingManager,
+                            val rideHailIterationHistory: RideHailIterationHistory,
+                            val routeHistory: RouteHistory,
+                            val beamSkimmer: BeamSkimmer,
+                            val travelTimeObserved: TravelTimeObserved,
+                            val geo: GeoUtils,
+                            val networkHelper: NetworkHelper
 ) extends Mobsim
     with LazyLogging
     with MetricsSupport {
@@ -115,12 +115,12 @@ class BeamMobsim @Inject()(
 }
 
 class BeamMobsimIteration(
-  val beamServices: BeamServices,
-  val rideHailSurgePricingManager: RideHailSurgePricingManager,
-  val rideHailIterationHistory: RideHailIterationHistory,
-  val routeHistory: RouteHistory,
-  val beamSkimmer: BeamSkimmer,
-  val travelTimeObserved: TravelTimeObserved
+                           val beamServices: BeamServices,
+                           val rideHailSurgePricingManager: AdaptiveRideHailSurgePricingManager,
+                           val rideHailIterationHistory: RideHailIterationHistory,
+                           val routeHistory: RouteHistory,
+                           val beamSkimmer: BeamSkimmer,
+                           val travelTimeObserved: TravelTimeObserved
 ) extends Actor
     with ActorLogging
     with MetricsSupport {
